Dev 3 rework expense data input (#11)
* [DEV-3] add expense_data_input parameter

* [DEV-3] add_expense_input parser depends on layout klass

* [DEV-3] remove log_factory

* [DEV-3] add default layout factory

* [DEV-3] parser_base

* [DEV-3]
- add unknonwn factory error
- add needed factories options

* [DEV-3] parser base methods

* [DEV-3] split default and named inputs params parsers

* [DEV-3] add expense via named inputs

* [DEV-3] add expense range to spreadsheet

* [DEV-3] remove params builder stub

* [DEV-3] named parameters

* [DEV-3] rubocop

+++ app/services/telegram/messages/layouts/spreadsheets/layouts/delete.rb
@@ -6,18 +6,30 @@ module Telegram
       module Spreadsheets
         module Layouts
           class Delete < Base
-            string :document_id
+            string :document_id, default: nil
 
-            define_action(:delete_table)
+            define_action(:enter_spreadsheets_params, 'Удалить')
 
             private
 
-            def delete_table
+            def enter_spreadsheets_params
+              unless valid_spreadsheets_params?
+                messages << 'Невалидные данные таблицы'
+                return handle_messages do
+                  layouts_factory(layout_name: :list_tables).run!(bot: bot, user: user, action_name: :list_all_actions)
+                end
+              end
+
               spreadsheet.destroy
+              messages << 'Таблица удалена'
 
               messages << layouts_factory(layout_name: :index)
                           .run!(bot: bot, user: user, action_name: :list_all_actions)
               messages.flatten!
+            end
+
+            def valid_spreadsheets_params?
+              document_id.present? && spreadsheet.present?
             end
 
             def spreadsheet
